docs: behavior map comp
